Add promtool check and test in the CI
This PR adds the check and test of Prometheus rules. Because
the exporter is going to change and also the metrics/rules, the
only unit test covered is for if the scrape is up.

When the new exporter is used, then the new rules and unit tests
will be implemented.