fix: Add xml coverage generation and fix tics workflow (#193)
### Workflow updates:

*
[`.github/workflows/tics_run_sh_ghaction_test.yml`](diffhunk://#diff-594936f70c4be3041fe643be75996fd4d5d1c50675521239e0d7811b85790b90L10-R17):
Updated the `runs-on` parameter to include `jammy` and `tiobe` labels
for better compatibility with the target environment.
*
[`.github/workflows/tics_run_sh_ghaction_test.yml`](diffhunk://#diff-594936f70c4be3041fe643be75996fd4d5d1c50675521239e0d7811b85790b90L34-R35):
Changed the folder name for coverage results from `.cover` to `cover`.

### Dependency installation:

*
[`.github/workflows/tics_run_sh_ghaction_test.yml`](diffhunk://#diff-594936f70c4be3041fe643be75996fd4d5d1c50675521239e0d7811b85790b90L10-R17):
Modified the system dependencies installation command to install
`python3-venv` instead of `python3.10-venv` for broader Python version
support.

### Test coverage reporting:

*
[`tox.ini`](diffhunk://#diff-ef2cef9f88b4fe09ca3082140e67f5ad34fb65fb6e228f119d3812261ae51449R56):
Added a command to generate an XML report for test coverage, enabling
integration with external tools.