[DPE-7002] Update opensearch health checks (#215)
## Description of issue or feature:
This PR addresses the issue mentioned in #177. After initial discussion,
it was decided that the opensearch-dashboard charm should stay in a
blocked status if the opensearch charm is not healthy or unreachable.
This is due to the tight coupling between the two charms. The user is
therefore asked to make interventions if the opensearch charm is not in
a healthy state. At the current state, we don't make difference between
opensearch being sel-healed or restarting and the case of opensearch not
working correctly.

Therefore this PR makes improvement over the opensearch health checks.

## Solution:
- Remove the `app_healthy` function and make calls directly to
opensearch_ok to avoid confusions.
- Add tenacity to the health checks. 
- Log the exception happening when json decode error occurs.
- Update charm status message when opensearch is in a yellow state.

## How was this change tested?
- [X] Manually
- [X] Unit tests
- [X] Integration tests

---------

Co-authored-by: akram09 <akram.boutouchent@eurecom.fr>